Move hex block rotate tests to unique file
